The .alp file that will house the model.
Added our groups contributions from our meeting on Feb 19. We added parameters to the model along with two agents. A resident agent which will be the representation of a care home resident within our model. We also added the infectious behaviour agent which consists of a SEIR state chart which will represent the infection. The SEIR statechart may need some modifying as it will only work when placed in the resident agent.